requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
distro==1.9.0
Django==5.1.2
django-ledger==0.6.4
django-treebeard==4.7.1
Faker==30.8.0
graphene==3.4
graphene-django==3.2.2
graphql-core==3.2.5
graphql-relay==3.2.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
Markdown==3.7
numpy==2.1.2
oauth2_provider==0.0
ofxtools==0.9.5
openai==1.52.2
pandas==2.2.3
pillow==11.0.0
promise==2.3
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
text-unidecode==1.3
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
